FEATURE: Peek at pastes
Allows previewing of pastes in the TUI.
By pressing `Space` on a paste in a list, it shows the contents of the
paste in the TUI.

If a paste has multiple files, pressing `Ctrl-n` and `Ctrl-p` will
switch between those files.

Pressing `q`, `Esc`, or `Ctrl-c` will navigate back to the list.